Major change: Choices to User
I changed the script significantly, giving the user the option to get either the spikeshape files one by one or all at once (in case he/she wants to leave it running for a long time). Also put the option to plot or not an example of each spikeshape file.